third_party/blink/web_tests/transitions/font-family-during-transition-expected.txt

PASS - "font-size" property for "box" element at 1s saw something close to: 15
PASS: width is close to control
